# api-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FutLendas API",
    "version": "1.0.0",
    "description": "Documentação da API backend para o aplicativo FutLendas."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Servidor de Desenvolvimento Local" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
...
      "Jogador": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "nome": { "type": "string", "example": "Lionel Messi" },
          "role": { "type": "string", "default": "player", "example": "player" },
          "joga_recuado": { "type": "boolean", "default": false, "example": false },
          "nivel": { "type": "integer", "default": 1, "example": 5 },
          "posicao": {
            "type": "string",
            "enum": ["linha", "goleiro"],
            "default": "linha",
            "example": "linha"
          },
          "foto_url": {
            "type": "string",
            "nullable": true,
            "format": "url",
            "example": "http://example.com/foto.jpg"
          }
        }
      },
...
      "TimeFixo": {
        "type": "object",
        "description": "Schema para times fixos (campeonato)",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "nome": { "type": "string", "example": "Barcelona" },
          "logo_url": { "type": "string", "format": "url", "nullable": true }
        }
      },
...
      "RodadaSaveTimesInput": {
        "type": "object",
        "required": ["times"],
        "properties": {
          "times": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "object",
              "properties": {
                "nome": { "type": "string", "example": "Time Amarelo" },
                "jogadores": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["id"],
                    "properties": { "id": { "type": "integer", "example": 1 } }
                  }
                }
              }
            }
          }
        }
      },
...
      "TimeFixoInput": {
        "type": "object",
        "required": ["nome"],
        "properties": {
          "nome": { "type": "string", "example": "Real Madrid" },
          "logo_url": { "type": "string", "format": "url", "nullable": true }
        }
      },
      "TimeFixoUpdateInput": {
        "type": "object",
        "minProperties": 1,
        "properties": {
          "nome": { "type": "string", "example": "Real Madrid CF" },
          "logo_url": { "type": "string", "format": "url", "nullable": true }
        }
      },
      "TimeAddJogadoresInput": {
        "type": "object",
        "required": ["jogador_ids"],
        "properties": {
          "jogador_ids": {
            "type": "array",
            "items": { "type": "integer" },
            "minItems": 1,
            "example": [1, 5, 10]
          }
        }
      },
...
      "CampeonatoRegisterTimeInput": {
        "type": "object",
        "required": ["time_id"],
        "properties": { "time_id": { "type": "integer" } }
      },
...
    },
...
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/campeonatos/{campeonato_id}/times": {
      "get": {
        "summary": "Listar times do campeonato",
        "description": "Retorna todos os times que estão inscritos no campeonato especificado.",
        "tags": ["Campeonatos"],
        "parameters": [
          {
            "in": "path",
            "name": "campeonato_id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": { "200": { "description": "Lista de times retornada com sucesso." } }
      },
      "post": {
        "summary": "Registrar time no campeonato (Admin)",
        "description": "Inscreve um time existente em um campeonato que está na fase de 'inscricao'.",
        "tags": ["Campeonatos"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "campeonato_id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CampeonatoRegisterTimeInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "Time registrado com sucesso." },
          "400": { "description": "Dados inválidos." },
          "401": { "description": "Não autenticado." },
          "403": { "description": "Acesso negado." },
          "409": { "description": "Conflito: O time já está registrado neste campeonato." }
        }
      }
    },
...
    "/rodadas/{rodada_id}/times": {
      "get": {
        "summary": "Listar times de uma rodada",
        "description": "Retorna os times que foram sorteados e salvos para a rodada, com seus respectivos jogadores.",
        "tags": ["Rodadas"],
        "parameters": [
          { "in": "path", "name": "rodada_id", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Lista de times da rodada." } }
      },
      "post": {
        "summary": "Salvar times da rodada (Admin)",
        "description": "Define e salva a composição dos times para a rodada. Esta ação substitui qualquer sorteio de times salvo anteriormente para esta rodada.",
        "tags": ["Rodadas"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "rodada_id", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RodadaSaveTimesInput" }
            }
          }
        },
        "responses": {
          "200": { "description": "Times salvos com sucesso." },
          "400": { "description": "Dados inválidos." },
          "401": { "description": "Não autenticado." },
          "403": { "description": "Acesso negado." },
          "404": { "description": "Rodada não encontrada." }
        }
      }
    },
...
    "/times": {
      "get": {
        "summary": "Listar todos os times fixos",
        "description": "Retorna a lista completa de times fixos cadastrados no sistema.",
        "tags": ["Times"],
        "responses": {
          "200": {
            "description": "Lista de times retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/TimeFixo" } }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar novo time fixo (Admin)",
        "description": "Cadastra um novo time fixo no sistema, que pode ser usado em campeonatos.",
        "tags": ["Times"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/TimeFixoInput" } }
          }
        },
        "responses": {
          "201": { "description": "Time criado com sucesso." },
          "400": { "description": "Dados inválidos (ex: nome faltando)." },
          "401": { "description": "Não autenticado." },
          "403": { "description": "Acesso negado." },
          "409": { "description": "Conflito: Um time com este nome já existe." }
        }
      }
    },
    "/times/{id}": {
      "get": {
        "summary": "Buscar time por ID",
        "description": "Retorna os detalhes de um time fixo específico.",
        "tags": ["Times"],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Dados do time retornados com sucesso.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/TimeFixo" } }
            }
          },
          "404": { "description": "Time não encontrado." }
        }
      },
      "put": {
        "summary": "Atualizar time (Admin)",
        "description": "Atualiza o nome e/ou a URL do logo de um time existente.",
        "tags": ["Times"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/TimeFixoUpdateInput" } }
          }
        },
        "responses": {
          "200": { "description": "Time atualizado com sucesso." },
          "400": { "description": "Dados inválidos." },
          "401": { "description": "Não autenticado." },
          "403": { "description": "Acesso negado." },
          "404": { "description": "Time não encontrado." },
          "409": { "description": "Conflito: Já existe outro time com o nome fornecido." }
        }
      }
    },
    "/times/{id}/jogadores": {
      "get": {
        "summary": "Listar jogadores de um time",
        "description": "Retorna la lista de todos os jogadores que pertencem ao elenco de um time fixo.",
        "tags": ["Times"],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Lista de jogadores do time retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Jogador" } }
              }
            }
          },
          "404": { "description": "Time não encontrado." }
        }
      },
      "post": {
        "summary": "Adicionar jogadores a um time (Admin)",
        "description": "Vincula um ou mais jogadores a um time. Se um jogador já estiver no time, ele é ignorado.",
        "tags": ["Times"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TimeAddJogadoresInput" }
            }
          }
        },
        "responses": {
          "200": { "description": "Jogadores adicionados com sucesso." },
          "400": { "description": "Dados inválidos (ex: lista de IDs vazia)." },
          "401": { "description": "Não autenticado." },
          "403": { "description": "Acesso negado." },
          "404": { "description": "Time não encontrado." }
        }
      }
    },
    "/times/{id}/jogadores/{jogador_id}": {
      "delete": {
        "summary": "Remover jogador de um time (Admin)",
        "description": "Remove a associação de um jogador específico de um time.",
        "tags": ["Times"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID do time."
          },
          {
            "in": "path",
            "name": "jogador_id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID do jogador a ser removido."
          }
        ],
        "responses": {
          "200": { "description": "Jogador removido do time com sucesso." },
          "401": { "description": "Não autenticado." },
          "403": { "description": "Acesso negado." },
          "404": { "description": "A associação entre time e jogador não foi encontrada." }
        }
      }
    },
...
  },
...
}